support multi-channel gray slide
add two functions:

```
void openslide_read_region_gray8(openslide_t *osr, uint8_t *dest,
                                 int64_t x, int64_t y, int32_t level,
                                 int64_t w, int64_t h);

void openslide_read_region_gray16(openslide_t *osr, uint8_t *dest,
                                  int64_t x, int64_t y, int32_t level,
                                  int64_t w, int64_t h);
```

They accept similar set of parameters as openslide_read_region(). The
only difference is the `level` parameter. Since `level` is the index
into `osr->levels` array, and `osr->levels` now stores `struct
_openslide_level` for all levels of all channels, caller needs calculate
`level` as:

    num_of_channels * num_of_level_in_one_channel + level_in_channel

+++ src/openslide-decode-jxr.h
@@ -25,7 +25,7 @@
 #include <stdint.h>
 
 
-bool _openslide_jxr_decode_buf(const void *src, int64_t src_len, uint32_t *dst,
+bool _openslide_jxr_decode_buf(const void *src, int64_t src_len, uint8_t *dst,
                                int64_t dst_len, GError **err);
 
 bool _openslide_jxr_dim(const void *data, size_t data_len, uint32_t *width,
